remove need of nodeport from the 'weird username tests' for bridge
