+++ systemtest/src/test/java/io/strimzi/systemtest/bridge/HttpBridgeTlsST.java
@@ -5,9 +5,15 @@
 package io.strimzi.systemtest.bridge;
 
 import io.strimzi.api.kafka.model.bridge.KafkaBridgeResources;
+import io.strimzi.api.kafka.model.bridge.KafkaBridgeSpec;
+import io.strimzi.api.kafka.model.bridge.KafkaBridgeSpecBuilder;
 import io.strimzi.api.kafka.model.common.CertSecretSource;
+import io.strimzi.api.kafka.model.common.PasswordSecretSource;
 import io.strimzi.api.kafka.model.kafka.KafkaResources;
 import io.strimzi.api.kafka.model.kafka.listener.GenericKafkaListenerBuilder;
+import io.strimzi.api.kafka.model.kafka.listener.KafkaListenerAuthentication;
+import io.strimzi.api.kafka.model.kafka.listener.KafkaListenerAuthenticationScramSha512;
+import io.strimzi.api.kafka.model.kafka.listener.KafkaListenerAuthenticationTls;
 import io.strimzi.api.kafka.model.kafka.listener.KafkaListenerType;
 import io.strimzi.api.kafka.model.user.KafkaUser;
 import io.strimzi.systemtest.AbstractST;
@@ -17,6 +23,7 @@ import io.strimzi.systemtest.annotations.ParallelTest;
 import io.strimzi.systemtest.kafkaclients.internalClients.BridgeClients;
 import io.strimzi.systemtest.kafkaclients.internalClients.BridgeClientsBuilder;
 import io.strimzi.systemtest.kafkaclients.internalClients.KafkaClients;
+import io.strimzi.systemtest.kafkaclients.internalClients.KafkaClientsBuilder;
 import io.strimzi.systemtest.resources.NodePoolsConverter;
 import io.strimzi.systemtest.resources.ResourceManager;
 import io.strimzi.systemtest.storage.TestStorage;
@@ -90,13 +97,167 @@ class HttpBridgeTlsST extends AbstractST {
         ClientUtils.waitForClientsSuccess(testStorage.getProducerName(), testStorage.getConsumerName(), testStorage.getNamespaceName(), testStorage.getMessageCount());
     }
 
+    @ParallelTest
+    void testTlsScramShaAuthWithWeirdUsername() {
+        final TestStorage testStorage = new TestStorage(ResourceManager.getTestContext());
+        // Create weird named user with . and more than 64 chars -> SCRAM-SHA
+        final String weirdUserName = "jjglmahyijoambryleyxjjglmahy.ijoambryleyxjjglmahyijoambryleyxasd.asdasidioiqweioqiweooioqieioqieoqieooi";
+
+        // Initialize PasswordSecret to set this as PasswordSecret in MirrorMaker spec
+        final PasswordSecretSource passwordSecret = new PasswordSecretSource();
+        passwordSecret.setSecretName(weirdUserName);
+        passwordSecret.setPassword("password");
+
+        // Initialize CertSecretSource with certificate and secret names for consumer
+        CertSecretSource certSecret = new CertSecretSource();
+        certSecret.setCertificate("ca.crt");
+        certSecret.setSecretName(KafkaResources.clusterCaCertificateSecretName(testStorage.getClusterName()));
+
+        KafkaBridgeSpec bridgeSpec = new KafkaBridgeSpecBuilder()
+            .withNewKafkaClientAuthenticationScramSha512()
+                .withUsername(weirdUserName)
+                .withPasswordSecret(passwordSecret)
+            .endKafkaClientAuthenticationScramSha512()
+            .withNewTls()
+                .withTrustedCertificates(certSecret)
+            .endTls()
+            .build();
+
+        testWeirdUsername(weirdUserName, new KafkaListenerAuthenticationScramSha512(), bridgeSpec, testStorage);
+    }
+
+    @ParallelTest
+    void testTlsAuthWithWeirdUsername() {
+        final TestStorage testStorage = new TestStorage(ResourceManager.getTestContext());
+        // Create weird named user with . and maximum of 64 chars -> TLS
+        final String weirdUserName = "jjglmahyijoambryleyxjjglmahy.ijoambryleyxjjglmahyijoambryleyxasd";
+
+        // Initialize CertSecretSource with certificate and secret names for consumer
+        CertSecretSource certSecret = new CertSecretSource();
+        certSecret.setCertificate("ca.crt");
+        certSecret.setSecretName(KafkaResources.clusterCaCertificateSecretName(testStorage.getClusterName()));
+
+        KafkaBridgeSpec bridgeSpec = new KafkaBridgeSpecBuilder()
+            .withNewKafkaClientAuthenticationTls()
+                .withNewCertificateAndKey()
+                    .withSecretName(weirdUserName)
+                    .withCertificate("user.crt")
+                    .withKey("user.key")
+                .endCertificateAndKey()
+            .endKafkaClientAuthenticationTls()
+            .withNewTls()
+                .withTrustedCertificates(certSecret)
+            .endTls()
+            .build();
+
+        testWeirdUsername(weirdUserName, new KafkaListenerAuthenticationTls(), bridgeSpec, testStorage);
+    }
+
+    @SuppressWarnings({"checkstyle:MethodLength"})
+    private void testWeirdUsername(String weirdUserName, KafkaListenerAuthentication auth,
+                                   KafkaBridgeSpec spec, TestStorage testStorage) {
+        String bridgeProducerName = testStorage.getProducerName() + "-" + TestConstants.BRIDGE;
+        String bridgeConsumerName = testStorage.getConsumerName() + "-" + TestConstants.BRIDGE;
+
+        resourceManager.createResourceWithWait(
+            NodePoolsConverter.convertNodePoolsIfNeeded(
+                KafkaNodePoolTemplates.brokerPool(testStorage.getNamespaceName(), testStorage.getBrokerPoolName(), testStorage.getClusterName(), 3).build(),
+                KafkaNodePoolTemplates.controllerPool(testStorage.getNamespaceName(), testStorage.getControllerPoolName(), testStorage.getClusterName(), 1).build()
+            )
+        );
+        resourceManager.createResourceWithWait(KafkaTemplates.kafkaEphemeral(testStorage.getClusterName(), 3, 1)
+            .editMetadata()
+                .withNamespace(testStorage.getNamespaceName())
+            .endMetadata()
+            .editSpec()
+                .editKafka()
+                    .withListeners(new GenericKafkaListenerBuilder()
+                        .withName(TestConstants.TLS_LISTENER_DEFAULT_NAME)
+                        .withPort(9093)
+                        .withType(KafkaListenerType.INTERNAL)
+                        .withTls(true)
+                        .withAuth(auth)
+                        .build()
+                    )
+                .endKafka()
+            .endSpec()
+            .build());
+
+        BridgeClients kafkaBridgeClientJob = new BridgeClientsBuilder()
+            .withProducerName(bridgeProducerName)
+            .withConsumerName(bridgeConsumerName)
+            .withBootstrapAddress(KafkaBridgeResources.serviceName(testStorage.getClusterName()))
+            .withComponentName(KafkaBridgeResources.componentName(testStorage.getClusterName()))
+            .withTopicName(testStorage.getTopicName())
+            .withMessageCount(testStorage.getMessageCount())
+            .withPort(TestConstants.HTTP_BRIDGE_DEFAULT_PORT)
+            .withNamespaceName(testStorage.getNamespaceName())
+            .build();
+
+        // Create topic
+        resourceManager.createResourceWithWait(KafkaTopicTemplates.topic(testStorage).build());
+
+        // Create user
+        if (auth.getType().equals(TestConstants.TLS_LISTENER_DEFAULT_NAME)) {
+            resourceManager.createResourceWithWait(KafkaUserTemplates.tlsUser(testStorage.getNamespaceName(), testStorage.getClusterName(), weirdUserName)
+                .editMetadata()
+                    .withNamespace(testStorage.getNamespaceName())
+                .endMetadata()
+                .build());
+        } else {
+            resourceManager.createResourceWithWait(KafkaUserTemplates.scramShaUser(testStorage.getNamespaceName(), testStorage.getClusterName(), weirdUserName)
+                .editMetadata()
+                    .withNamespace(testStorage.getNamespaceName())
+                .endMetadata()
+                .build());
+        }
+
+        // Deploy http bridge
+        resourceManager.createResourceWithWait(KafkaBridgeTemplates.kafkaBridge(testStorage.getClusterName(), KafkaResources.tlsBootstrapAddress(testStorage.getClusterName()), 1)
+            .editMetadata()
+                .withNamespace(testStorage.getNamespaceName())
+            .endMetadata()
+            .withNewSpecLike(spec)
+                .withBootstrapServers(KafkaResources.tlsBootstrapAddress(testStorage.getClusterName()))
+                .withNewHttp(TestConstants.HTTP_BRIDGE_DEFAULT_PORT)
+                .withNewConsumer()
+                    .addToConfig(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest")
+                .endConsumer()
+            .endSpec()
+            .build());
+
+        resourceManager.createResourceWithWait(kafkaBridgeClientJob.consumerStrimziBridge());
+
+        final KafkaClients kafkaClients = new KafkaClientsBuilder()
+            .withNamespaceName(testStorage.getNamespaceName())
+            .withBootstrapAddress(KafkaResources.tlsBootstrapAddress(testStorage.getClusterName()))
+            .withUsername(weirdUserName)
+            .withMessageCount(testStorage.getMessageCount())
+            .withTopicName(testStorage.getTopicName())
+            .withProducerName(testStorage.getProducerName())
+            .withConsumerName(testStorage.getConsumerName())
+            .build();
+
+        if (auth.getType().equals(TestConstants.TLS_LISTENER_DEFAULT_NAME)) {
+            // tls producer
+            resourceManager.createResourceWithWait(kafkaClients.producerTlsStrimzi(testStorage.getClusterName()));
+        } else {
+            // scram-sha producer
+            resourceManager.createResourceWithWait(kafkaClients.producerScramShaTlsStrimzi(testStorage.getClusterName()));
+        }
+
+        ClientUtils.waitForClientSuccess(testStorage.getProducerName(), testStorage.getNamespaceName(), testStorage.getMessageCount());
+
+        ClientUtils.waitForClientSuccess(bridgeConsumerName, testStorage.getNamespaceName(), testStorage.getMessageCount());
+    }
+
     @BeforeAll
     void setUp() {
         suiteTestStorage = new TestStorage(ResourceManager.getTestContext());
 
         clusterOperator = clusterOperator.defaultInstallation()
-                .createInstallation()
-                .runInstallation();
+            .createInstallation()
+            .runInstallation();
 
         LOGGER.info("Deploying Kafka and KafkaBridge before tests");
